Remove a workaround for an old pygtk bug
The bug was documented as a "focus bug in pygtk" on nested Gtk.MenuItem
objects. The workaround was to switch the signal handler from "activate"
to "button-release-event". This was done in Mar 2012 and pygtk is no
longer supported upstream. I reverted the workaround and noticed no
issues during testing.

This allows all MenuItem callbacks to share a common interface and
should allow keyboard control of sub-menus.